refactor(snapshot): Remove MergeClassesTask, scan classpath in-place
Instead of merging all classes into an intermediate directory,
ExportPreviewMetadataTask now scans directories and JARs directly
from the classpath.
